megano/requirements.txt
asgiref==3.6.0
Django==4.2
django-debug-toolbar==4.0.0
django-filter==22.1
djangorestframework==3.14.0
Pillow==9.4.0
pytz==2022.7.1
sqlparse==0.4.2
drf_spectacular==0.28.0
python-decouple==3.8
celery==5.4.0
redis==5.2.1
ruff==0.11.2